+++ src/runtime/index.js
@@ -26,6 +26,7 @@ export { TeeNode } from './tee-node';
 export { HookNode } from './hook-node';
 export { CallbackNode } from './callback-node';
 export { EchoNode } from './echo-node';
+export { NullNode } from './null-node';
 export { TimerNode, GRID_PHASE_MS } from './timer-node';
 export { HeartbeatNode } from './heartbeat-node';
 export { CommandInterpreterNode } from './command-interpreter-node';